backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
python-dotenvfirebase-admin